requirements.txt
annotated-types==0.7.0
anyio==4.10.0
apify==2.7.3
apify_client==1.12.2
apify_fingerprint_datapoints==0.0.3
apify_shared==1.5.0
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
Brotli==1.1.0
browserforge==1.2.3
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
crawlee==0.6.12
cryptography==45.0.6
cssselect2==0.8.0
distro==1.9.0
Django==5.2.5
django-debug-toolbar==6.0.0
eval_type_backport==0.2.2
filelock==3.19.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
lazy-object-proxy==1.10.0
lxml==6.0.1
more-itertools==10.7.0
multidict==6.6.4
openai==1.99.9
openrouter==1.0
oscrypto==1.3.0
packaging==25.0
pillow==11.3.0
propcache==0.3.2
Protego==0.5.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.6.1
pydantic_core==2.33.2
pyee==13.0.0
pyHanko==0.30.0
pyhanko-certvalidator==0.28.0
pyparsing==3.2.3
pypdf==6.0.0
python-bidi==0.6.6
python-dotenv==1.1.1
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.4
requests-file==2.1.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sortedcollections==2.1.0
sortedcontainers==2.4.0
sqlparse==0.5.3
stripe==12.4.0
svglib==1.5.1
tinycss2==1.4.0
tldextract==5.3.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
websockets==15.0.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
yarl==1.20.1
zstandard==0.23.0
django-allauth==65.11.2
psycopg2-binary==2.9.10
django-encrypted-model-fields==0.6.5
